+++ src/main/java/racingcar/view/io/InputView.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.view.io;
 
 import java.util.Scanner;
